invertedai_drive/requirements.txt
torch==1.11.0
python-dotenv==0.20.0
numpy==1.21
